requirements.txt
autopep8==2.0.2
greenlet==2.0.2
# Editable Git install with no remote (insert-database==0.1)
-e /home/lionheart/dev/python_insert_db
numpy==1.24.3
pandas==2.0.1
psycopg2==2.9.6
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.12
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
